Add functions to create MPO Hamiltonian operators
Some functions were moved from the `tvdp` branch of
`tebd_pseudomodes_ttedopa` to this package, so that they can be used
in every script without copying them every time. They create some
Hamiltonian and Lindbladian operators in an OpSum form, which then can
be feeded into the MPO constructor to create operators suitable to the
TDVP solution of the equations of motion.