Add default discretization for OrthoPoly
Add a default `discretization` keyword argument for the OrthoPoly when
computing the TEDOPA chain coefficients. This allows the end user to
avoid specifying it when calling `chainmapcoefficients`, something that
also forced them to include "using PolyChaos" in each script since
otherwise the symbols for the possible discretization arguments were not
defined.