Move and unify vectorized S=1/2 def. in new file
Define a new vectorized 1/2-spin SiteType, "vS=1/2", which
replaces both the "vecS=1/2" and "HvS=1/2" types. The new type uses the
Gell-Mann basis for the vectorization procedure, since there is
probably no advantage in using the canonical basis. Moreover, the
definitions of states and operators were rewritten so that they make the
most of the pre-existing ITensors functions, in order to have as few
new functions as possible. The new definitions are found in a new
file, named site_types/vectorized_spinhalf.jl.

Some old functions were deprecated, but all previous OpNames, StateNames
have been "rewired" so that they call the appropiate new ones; the two
"vecS=1/2" and "HvS=1/2" types are now aliases of the new "vS=1/2" type.