Fix aliasing of "HvS=1/2" and "vecS=1/2" SiteTypes
Fix the `ITensors.state` function with the aliases so that it reproduces
exactly the signature of the `ITensors.state` definitions for the
"vS=1/2" SiteType (the same goes for `ITensors.op`).
Previously, there was an additional Index argument which was not present
in the various state definitions, which as a consequence were never seen
by the `state(::Index, ::String)` function from ITensor.

The same, luckily, does not happen for the "vOsc" SiteType since in that
case we define another "intermediate" function that adds the dimension
of the Index to the arguments; it is precisely this function that does
the required redirection for the aliased types.