Remove explicit definition of Up and Dn states
Use the "Up" and "Dn" state vectors provided by ITensors, instead of
defining them explicitly. This way the definitions of states and
operators automatically match, without the need to check that, e.g. S+ *
Up = 0 (one could switch the definitions of Up and Dn states, but then
every operator must adhere to the chosen order of the basis in C^2).