Use one fewer coupling constant in spin_chain
Use an array of coupling constants that has one fewer element than the
array of frequencies (and enforce this with an @assert statement).
This eliminates the need to start counting the coupling constants from
the second element, which was a behaviour specific to the simulation
scripts and not a general one.