Add new `vFermion` site type
Add a new vectorized site type `vFermion`, corresponding to the
`Fermion` type provided by ITensors. This actually replaces the `vS=1/2`
site type, which now has a more precise meaning, namely the vectorized
form of a `S=1/2` type.
This duplication allows defining a version of the various operators such
as `spin_chain` and so on that does not include Pauli strings from the
Jordan-Wigner transformation, precisely the one used by the `vS=1/2`
type. The `vFermion` type continues to use the Jordan-Wigner
transformation.