Add automatic postmult/premult for `vFermion` type
Add a new op(::OpName, ::SiteType"vFermion"; kwargs...) method which
parses the operator name for `\cdot`s, and constructs the post- or
pre-multiplication operator for vectorized types accordingly.
This automatically extends all "Fermion" operators to "vFermion"
operators, acting on the left or on the right by multiplication, without
having to define them manually.