Use two site type args in `exchange_interaction`
Redefine the `exchange_interaction` methods to accept two site types
instead of one. This allows more flexibility in the definitions (i.e. we
may introduce an exchange interaction between `Electron` and `Fermion`
site types).
The site types are, as always, deduced automatically from the tags of
the `Index` variables supplied to the `exchange_interaction` function,
so this commit does not break compatibility with older scripts.