Use real coupling constants in exchange inter. op
There is no standardized way (that I know of) to choose if, given a
complex coupling constant `a` and annihilation/creation operators `c_1`,
`cdag_1`, `c_2` and `cdag_2` on two different sites, the
exchange-interaction operator should be written as
  `a cdag_1 c_2 + conj(a) cdag_2 c_1`
or
  `conj(a) cdag_1 c_2 + a cdag_2 c_1`
just to give a simple example.
We could just tell the user which one between `a` and `conj(a)` they
should use if they want to construct the operator in the desired way,
but it seems too complicated, so until a better implementation is found
we use real coupling constants (this is probably enough for all
real-life cases after all), forcing this behaviour by annotating the
type (i.e. Real) on the coupling constant keyword arguments.