Fix wrong `vFermion` states
The `vFermion` definitions, especially for the vectorized states, used
states and operators from both `Fermion` and `S=1/2` site types. As a
result, there was confusion and it caused some bugs, e.g. the `vFermion`
"Emp" state was based on the `S=1/2` "Dn" state, but the `a` operator
for `Fermion`s, added in 3484711, was based on the sigma^+ Pauli matrix.
States and operators were not written using the same basis, and
consequently some operators stopped working as expected when in commit
1f3255c the `σ-`/`σ+` operators were replaced by `a`/`a†` in more
complex constructions such as the closure operators.